HW3 . two tasks
